alacritty: remove `open` directive in `mouse`
When opening URLs, if the `open` directive is set, alacritty will
attempt to open the URL with whatever it is set to and has no fallback.
This causes issues on platforms other than Linux, as I had it set to
`xdg-open`. By deleting the directive, it defaults to whatever the
platform default is, restoring behaviour on alternate platforms.