mutt: Improve handling of notmuch mailboxes
I've swapped to using notmuch to handle my mail, which features a set of
awesomely complicated hooks to ensure that everything is in the right
IMAP folder for Gmail. These changes just help me ensure that my reading
of email in neomutt does what I intend it to do, instead of potentially
messing with IMAP flags and causing issues with notmuch.