notmuch: Improve post-new hook
Before, any email that was in the "Send Mail" folder would be removed
from the inbox, which technically isn't correct. Instead, we now just
mark it as sent, and only remove things that aren't in an inbox from
being tagged as "inbox".

This should allow for messages to be sent to self and for them to appear
in the inbox.