waybar: Add temperature to bar
On this Thinkpad E485 (which I do most of my work on - and all of my
wayland/sway related stuff), temperature is read from hwmon2, which maps
to k10temp.
K10temp provides the temperature of a Ryzen CPU (this has a Ryzen 5
2500U), and as the GPU is on the same package, conveniently both AMDGPU
(hwmon6) and the Thinkpad temperature (hwmon5) map to lower resolution
versions of the same temperature.