notmuch: Make pre-new hook POSIX compliant
It was previously not as $'' syntax isn't POSIX. Instead I've moved to
using `read -r`, which is POSIX compliant.

A typical issue with `read -r` is that if the final line doesn't end
with a newline it'll exit early, but as we're pushing a stream into it
from program output, we know that our output will end in a newline.