notmuch: Fix '&& ||' and quoting issues in pre-new
According to shellcheck SC2015, A && B || C can run C if A is true but B
is false. This wasn't too much of a problem for my scripts (C was just
"true"), but for cleanness sake, it probably makes sense to fix it.

Some of the `rm`s and conditionals relied on their parameters being
single words. This was generally true due to the file format of mbsync
mail entries, but for correctness sake this should be changed.