fontconfig: Use Noto Mono instead of Space Mono
I've been using Space Mono for close to a year and a half, just because
it has italic and bold support (which is quite rare for monospaced
fonts). I found out recently that Noto Mono also supports italics
(although it seems to be unofficial - it works in alacritty though) and
bold.

Noto Mono also doesn't have the same issues with ligatures as Space Mono
has, so swapping to it allows me to maintain a consistent font across
all my programs without needing to disable ligatures.

Another major boon of Noto Mono is that it is ubiquitous - I already
install all of the other Noto fonts when starting a new computer just
because having CJK characters and emoji support is most easily solved by
grabbing the Noto fonts.

(Funnily enough, a few years ago I had a bug with Noto Emoji with dwm
where it would cause crashes when emojis were in the title bar of
programs - e.g. Firefox. This seems to have either been fixed, or sway
just doesn't have the same problems)