profile: Properly quote source locations
They should have all been quoted in case the variables expand to
something with a space in them, but I had forgotten (and it wasn't
causing any issues for me).

This is just a preemptive fix so that I don't potentially run into any
issues with this in the future.