env,profile: Wrap expansions with quotes and braces
This just fixes things if they expand to something with a space in it.
In general, I don't create directories (or even files) with spaces in
them, but they may be downloaded or uncompressed from the internet or
something.

(I don't think there's any rule stating that ${HOME} can't expand to
something with a space in it, but I haven't seen any systems that would
end up with a space. Perhaps if someone purposefully makes a username
with a space?)