meta: Change license to unlicense
I previously put my dotfiles under the MIT (which is seriously overkill
and probably a little too restrictive for dotfiles), but have swapped to
the unlicense.

The MIT requires that it be included in all copies of the licensed
source, which could (if taken to the extreme) mean that anyone basing
things off my dotfiles could have to display the MIT somewhere so that
they're compliant. The Unlicense follows the same language as the MIT,
but without requiring the display of licensing.