sway: Add better comments
I really need to make my dotfiles a little more self-documenting. A
major issue that I foresee is that all of the reasons that I give for
why things are the way they are will be buried in git logs instead of in
the files themselves, which might be a bit confusing.