sway: Add hotkey for moving windows to workspace centre
As I've been using mpv and floating windows more, the need for a hotkey
to do this has increased. I often float mpv streams and float the window
off into a corner and work on something else, so being able to reset the
windows back into the centre of the screen.