sway: Set caps lock to super
By default, colemak sets the caps lock key to be backspace. I tend to
actually use the backspace key instead of caps lock, so remapping it to
something more useful is probably a good idea. This - of course - leads
to the question of what it should be remapped to.

My options are primarily between the escape and super keys, with the
first being useful for neovim, and the second being more useful for
window managing tasks in sway. I eventually decided on super as I
commonly use it and it is not in a particularly ergonomic position for
my thumb, however I may decide to eventually remap escape instead (and
most likely will if I ever have to use a touchbar mac again).