aerc: Add aerc config to dotfiles
I've actually really enjoyed using aerc for the last few days (and
definitely found it nicer to conceptualise than neomutt), so these are
the dotfiles that I've been using.

It is worth keeping in mind that I have all of my mail going through
notmuch, and so I've rebound some of the keys that would otherwise be
useless (e.g. bindings to archive or delete).
Using notmuch certainly has some downsides, but overall the indexing and
searching power that it provides has been more useful than not - the
hooks (which you can find in .config/notmuch/hooks) just have to make
sure that everything is exactly as would be expected by a Maildir.