currency-convert: Add script to repository
currency-convert is a utility that uses currency.world to convert
currencies. Importantly - it caches the conversion rate for 5d before
updating, which helps not overload currency.world.

This helps a lot with a script that I'm writing that determines the
amount of money that I spend -- most of my money is in AUD, but some
expenses are in USD.