nixpkgs: Add darwin configuration to repository
Nix-darwin[0] is a way of managing a system configuration of a
darwin-based system through Nix. This allows me to declare the packages
that I want to use and have them automatically updated and work across
package version.

The trickiest thing about this so far has been the method with which
launchd agents (specifically those for goimapnotify) are handled.

[0] https://github.com/LnL7/nix-darwin