profile: Add profile for apollo
Apollo is my macOS machine that I got for/from work. It is named after
Apollo, the first bishop of Corinth. Corinth was a massive trade capital
in the first century AD, so it makes sense to me to name my work
machines after people from Corinth.

The actual content of the profile adds a correct configuration for the
NIX_PATH to the environment variables.