notmuch: Fix pre-new hook
This fixes handling both the `deleted` tag and any folders that contain
special characters.

When doing `tag_to_folder`, if a conversation was previously deleted, it
could not be copied out as it would be excluded from the results of
`notmuch search`.

When doing `folder_prune`, special characters would be interpreted
according to their regular expression meaning (e.g. `[]` is a set, `*`
is wildcard, etc.). This could cause issues with trying to delete from
folders with a `[Gmail]` prefix.